requirements.txt
azure-devops==7.1.0b4
openai==1.6.1
python-dotenv==1.0.0
flask==2.3.3
markdown==3.5
pygments==2.16.1
difflib-renderer==0.1.0
langgraph==0.0.19
langchain==0.0.335
langchain-openai==0.0.2.post1
pydantic==2.4.2
